requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
motor>=3.1.2
httpx>=0.24.0
python-dotenv>=1.0.0
pymongo>=4.3.3
bson>=0.5.10
google-generativeai>=0.3.2  # Updated to latest version for compatibility with current Gemini models
